Refactor Access Point and MQTT module: update Access Point to stop after saving settings
