--- access_control/src/mqtt/mqtt_topics.cpp
@@ -1,5 +1,0 @@
-#include "mqtt_topics.hpp"
-
-// Definiowanie tematów MQTT
-const char *mqtt_topics::rfid_topic = "smarthome/security/RFID/data";
-const char *mqtt_topics::pinpad_topic = "smarthome/security/pinpad/data";
